requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
aiohttp
pytest
pytest-asyncio